requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
pydantic
